refactor pdal::plang::Array
