+++ python/pdal/PyPipeline.hpp
@@ -47,7 +47,6 @@
 #undef isspace
 
 #define PY_ARRAY_UNIQUE_SYMBOL LIBPDALPYTHON_ARRAY_API
-#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 
 #include <numpy/arrayobject.h>
